+++ app/Http/Controllers/BinTypeController.php
@@ -3,9 +3,9 @@
 namespace App\Http\Controllers;
 
 use Exception;
-use Illuminate\Support\Facades\Log;
 use App\Models\BinType;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 use App\Repositories\BinTypeRepository;
 use App\Http\Requests\BinType\{CreateBinTypeRequest, UpdateBinTypeRequest};
 
@@ -25,10 +25,10 @@ class BinTypeController extends Controller
     {
         try {
             $this->binTypeRepository->store($request->only('bin_type'));
-            return response()->json(['status' => 'success', 'msg' => 'Bin Type saved successfully.']);
+            return response()->json(['status' => 'success', 'msg' => 'Bin type saved successfully.']);
         } catch (Exception $e) {
-            Log::error('Error saving Bin Type: ' . $e->getMessage());
-            return response()->json(['status' => 'false', 'msg' => 'An error occurred while saving the Bin Type.']);
+            Log::error('Error saving bin type: ' . $e->getMessage());
+            return response()->json(['status' => 'false', 'msg' => 'An error occurred while saving the bin type.']);
         }
     }
 
@@ -49,22 +49,26 @@ class BinTypeController extends Controller
         try {
 
             $this->binTypeRepository->update($request->only('bin_type'), $binType->id);
-            //  $BinType->update($request->validated()); // Use validated data
-            return response()->json(['status' => 'success', 'msg' => 'Bin Type updated successfully.']);
+            return response()->json(['status' => 'success', 'msg' => 'Bin type updated successfully.']);
         } catch (Exception $e) {
-            Log::error('Error updating Bin Type: ' . $e->getMessage());
-            return response()->json(['status' => 'false', 'msg' => 'An error occurred while updating the Bin Type.']);
+            Log::error('Error updating bin type: ' . $e->getMessage());
+            return response()->json(['status' => 'false', 'msg' => 'An error occurred while updating the bin type.']);
         }
     }
 
     public function destroy($id)
     {
         try {
-            $this->binTypeRepository->delete($id);
-            return response()->json(['status' => 'success', 'msg' => 'Bin Type deleted successfully.']);
+            $binType = $this->binTypeRepository->findOrFail($id);
+            if ($binType) {
+                $this->binTypeRepository->delete($id);
+                return response()->json(['status' => 'success', 'msg' => 'Bin type deleted successfully.']);
+            } else {
+                return response()->json(['status' => 'false', 'msg' => 'Bin type not found.']);
+            }            
         } catch (Exception $e) {
-            Log::error('Error deleting Bin Type: ' . $e->getMessage());
-            return response()->json(['status' => 'false', 'msg' => 'An error occurred while deleting the Bin Type.']);
+            Log::error('Error deleting bin type: ' . $e->getMessage());
+            return response()->json(['status' => 'false', 'msg' => 'An error occurred while deleting the bin type.']);
         }
     }
     public function getBinTypeDataTable(Request $request)
