+++ app/Http/Controllers/EndUseSegmentController.php
@@ -25,10 +25,10 @@ class EndUseSegmentController extends Controller
     {
         try {
             $endUseSegments = $this->endUseSegmentRepository->store($request->only('end_use_segment'));
-            return response()->json(['status' => 'success', 'msg' => 'End Use Segment saved successfully.', 'data' => $endUseSegments]);
+            return response()->json(['status' => 'success', 'msg' => 'End use segment saved successfully.', 'data' => $endUseSegments]);
         } catch (Exception $e) {
-            Log::error('Error saving End Use Segment: ' . $e->getMessage());
-            return response()->json(['status' => 'false', 'msg' => 'An error occurred while saving the End Use Segment.']);
+            Log::error('Error saving end use segment: ' . $e->getMessage());
+            return response()->json(['status' => 'false', 'msg' => 'An error occurred while saving the end use segment.']);
         }
     }
     public function show($id)
@@ -45,21 +45,26 @@ class EndUseSegmentController extends Controller
     {
         try {
             $this->endUseSegmentRepository->update($request->only('end_use_segment'), $endUseSegment->id);
-            return response()->json(['status' => 'success', 'msg' => 'End Use Segment updated successfully.']);
+            return response()->json(['status' => 'success', 'msg' => 'End use segment updated successfully.']);
         } catch (Exception $e) {
-            Log::error('Error updating End Use Segment: ' . $e->getMessage());
-            return response()->json(['status' => 'error', 'msg' => 'An error occurred while updating the End Use Segment.']);
+            Log::error('Error updating end use segment: ' . $e->getMessage());
+            return response()->json(['status' => 'error', 'msg' => 'An error occurred while updating the end use segment.']);
         }
     }
 
     public function destroy($id)
     {
         try {
-            $this->endUseSegmentRepository->delete($id);
-            return response()->json(['status' => 'success', 'msg' => 'End Use Segment deleted successfully.']);
+            $endUseSegment = $this->endUseSegmentRepository->findOrFail($id);
+            if ($endUseSegment) {
+                $this->endUseSegmentRepository->delete($id);
+                return response()->json(['status' => 'success', 'msg' => 'End use segment deleted successfully.']);
+            } else {
+                return response()->json(['status' => 'false', 'msg' => 'End use segment not found.']);
+            }
         } catch (Exception $e) {
-            Log::error('Error deleting End Use Segment: ' . $e->getMessage());
-            return response()->json(['status' => 'false', 'msg' => 'An error occurred while deleting the End Use Segment.']);
+            Log::error('Error deleting end use segment: ' . $e->getMessage());
+            return response()->json(['status' => 'false', 'msg' => 'An error occurred while deleting the end use segment.']);
         }
     }
     public function getEndUseSegementDataTableList(Request $request)
